requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic[email]==2.8.2
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
websockets==12.0  # Added for test_websocket.py